+++ KyhTestingStartingCase/ShopAdmin/Commands/Product.cs
@@ -1,6 +1,9 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using Humanizer;
+using Microsoft.EntityFrameworkCore;
 using ShopGeneral.Data;
 using System.Net;
+using System.Text;
+using System.Text.Json;
 
 namespace ShopAdmin.Commands
 {
@@ -24,6 +27,29 @@ namespace ShopAdmin.Commands
             WriteToFile(listOfProductsWithoutImage);
         }
 
+        public void Export()
+        {
+            var products = new List<ShopGeneral.Data.Product>();
+            var p1 = new ShopGeneral.Data.Product();
+            p1.Manufacturer = new Manufacturer();
+            var p2 = new ShopGeneral.Data.Product();
+            p2.Manufacturer = new Manufacturer();
+            p1.Name = "p1";
+            p1.BasePrice = 100;
+            p1.Manufacturer.Name = "Apple";
+            p2.Name = "p2";
+            p2.BasePrice = 50;
+            p1.Manufacturer.Name = "Samsung";
+            JsonSerializerOptions options = new JsonSerializerOptions();
+            options.WriteIndented = true;
+            products.Add(p1);
+            products.Add(p2);
+            var result = new List<string[]>();
+            var JSONString = "";
+            JSONString = CreateJsonString(products);
+            WriteToFilePricerunner(JSONString);
+        }
+
         public bool DoesImageExist(string url)
         {
             HttpWebRequest request = (HttpWebRequest)WebRequest.Create(url);
@@ -40,6 +66,47 @@ namespace ShopAdmin.Commands
                 return false;
             }
             
+        }
+        public string CreateJsonString(List<ShopGeneral.Data.Product> products)
+        {
+            var stringBuilder = new StringBuilder();
+            stringBuilder.Append("{\n\"products\":[");
+
+            foreach (var product in products)
+            {
+                product.Manufacturer = new Manufacturer();
+                product.Manufacturer.Name = "sdgdfhdgfg";
+                product.Category = new ShopGeneral.Data.Category();
+                product.Category.Name = "vcxbcvbcv";
+                product.ImageUrl = "www.google.se";
+
+                stringBuilder.Append("\n{");
+                stringBuilder.Append($"\n\"id\":{product.Id},\n");
+                stringBuilder.Append($"\"title\":\"{product.Name}\",\n");
+                stringBuilder.Append($"\"description\":\" \",\n");
+                stringBuilder.Append($"\"price\":{product.BasePrice},\n");
+                stringBuilder.Append($"\"discountPercentage\":0,\n");
+                stringBuilder.Append($"\"rating\":0,\n");
+                stringBuilder.Append($"\"stock\":0,\n");
+                stringBuilder.Append($"\"brand\":\"{product.Manufacturer.Name}\",\n");
+                stringBuilder.Append($"\"category\":\"{product.Category.Name}\",\n");
+                stringBuilder.Append($"\"images\":[{product.ImageUrl}]\n");
+                stringBuilder.Append("},");
+            }
+
+            stringBuilder.Append("\"total\": 100,\r\n  \"skip\": 0,\r\n  \"limit\": 30\n}");
+            return stringBuilder.ToString();
+        }
+
+        public void WriteToFilePricerunner(string result)
+        {
+            string path = ".\\outfiles\\pricerunner";
+            string today = DateTime.Today.ToString("yyyyMMdd");
+            string filePath = $"{path}\\{today}.txt";
+
+            if (!Directory.Exists(path)) { Directory.CreateDirectory(path); }
+
+            File.WriteAllText(filePath, result);
         }
 
         public void WriteToFile(List<string> listOfProductsWithoutImage)
