+++ KyhTestingStartingCase/ShopAdmin/Properties/launchSettings.json
@@ -2,7 +2,7 @@
   "profiles": {
     "ShopAdmin": {
       "commandName": "Project",
-      "commandLineArgs": "verifyimage verifyimagesofproducts"
+      "commandLineArgs": "category checkempty"
     }
   }
 }
